compact and better names
